pyproject.toml
[tool.poetry]
name = "shacl-generator"
version = "0.1.0"
description = "A playground for generating SHACL shapes from legal texts"
authors = ["Ben Glaeser <ben.glaeser@foerderfunke.org>"]
readme = "README.md"
packages = [{include = "shacl_generator"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.31.1"
rdflib = "^7.0.0"
pandas = "^2.2.0"
pyyaml = "^6.0.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
PyPDF2 = "^3.0.1"
pycryptodome = "^3.20.0"
pyshacl = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 